+++ js/components/form/strip-zeros.js
