+++ src/resources/views/components/form/input.blade.php
@@ -47,7 +47,7 @@
         <input @if ($id) id="{{ $id }}" @endif
                type="{{ $attributes->get('type', 'text') }}"
                x-ref="{{ $attributes->get('x-ref', $ref) }}"
-               @if ($stripLeadingZeros) x-data="tallstackui_formStripLeadingZeros(@js($ref))" @endif
+               @if ($stripZeros) x-data="tallstackui_formInputStripZeros(@js($ref))" @endif
                @if ($prefix || $suffix) autocomplete="{{ $attributes->get('autocomplete', 'off') }}" @endif
                {{ $attributes->class([
                     $personalize['input.base'],
